requirements.txt
absl-py==1.4.0
altair==5.1.1
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Automat==22.10.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.6.2
blis==0.7.9
boto3==1.26.155
botocore==1.29.155
bs4==0.0.1
cachetools==5.3.0
catalogue==2.0.8
category-encoders==2.6.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
confection==0.0.4
constantly==15.1.0
contourpy==1.0.7
cryptography==41.0.3
cssselect==1.2.0
cycler==0.10.0
cymem==2.0.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
emoji==2.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.2
fire==0.5.0
flatbuffers==23.3.3
fonttools==4.39.3
fqdn==1.5.1
fsspec==2023.6.0
gast==0.4.0
gensim==4.3.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-images-download==2.8.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.54.0
h11==0.14.0
h5py==3.8.0
huggingface-hub==0.15.1
humanfriendly==10.0
hyperlink==21.0.0
idna==2.10
imageio==2.31.1
importlib-metadata==6.8.0
imutils==0.5.4
incremental==22.10.0
install==1.3.5
ipykernel==6.22.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jovian==0.2.47
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.25.0
keras==2.12.0
keras-facenet==0.3.2
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.3
lemminflect==0.2.3
libclang==16.0.0
llvmlite==0.40.1
lxml==4.9.3
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.3
missingno==0.5.2
mistune==3.0.2
ml-dtypes==0.1.0
mlxtend==0.22.0
mpmath==1.3.0
mtcnn==0.1.1
murmurhash==1.0.9
mysql==0.0.3
mysql-connector-python==8.1.0
mysqlclient==2.2.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.57.1
numpy==1.24.4
oauthlib==3.2.2
onnxruntime==1.15.1
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
overrides==7.4.0
packaging==23.1
pandas==2.1.1
pandocfilters==1.5.0
parsel==1.8.1
parso==0.8.3
pathy==0.10.1
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
plotly==5.14.1
pooch==1.7.0
preshed==3.0.8
progressbar==2.5
prometheus-client==0.17.1
prompt-toolkit==3.0.38
Protego==0.3.0
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
PyDispatcher==2.0.7
pydot==1.4.2
Pygments==2.15.1
PyMatting==1.1.8
pyOpenSSL==23.2.0
pyparsing==2.4.7
pyreadline3==3.4.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic-bin==0.4.14
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.4
QtPy==2.4.1
queuelib==1.6.2
raceplotly==0.1.7
referencing==0.30.2
regex==2023.5.5
rembg==2.0.49
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.3
roboflow==1.0.9
rpds-py==0.10.6
rsa==4.9
s3transfer==0.6.1
sahi==0.11.14
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
scrapeops-scrapy-proxy-sdk==1.0
Scrapy==2.10.0
scrapy-rotating-proxies==0.6.2
seaborn==0.12.2
selenium==4.10.0
Send2Trash==1.8.2
service-identity==23.1.0
shapely==2.0.1
simple-image-download==0.5
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.4.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
statsmodels==0.13.5
streamlit==1.27.2
streamlit-option-menu==0.3.6
sympy==1.12
tenacity==8.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-intel==2.12.0
tensorflow-io==0.31.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
terminaltables==3.1.10
thinc==8.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.7.4
tinycss2==1.2.1
tldextract==3.4.4
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.1
trio-websocket==0.10.3
Twisted==22.10.0
twisted-iocpsupport==1.0.4
typer==0.7.0
types-python-dateutil==2.8.19.14
typing==3.7.4.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
ultralytics==8.0.119
uri-template==1.3.0
urllib3==1.26.15
uuid==1.30
validators==0.22.0
virtualenv==20.24.3
w3lib==2.1.2
wasabi==1.1.1
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==2.3.3
wget==3.2
widgetsnbextension==4.0.7
wordcloud==1.9.2
wrapt==1.14.1
wsproto==1.2.0
xgboost==1.7.5
yellowbrick==1.5
yolov5==7.0.12
zipp==3.17.0
zope.interface==6.0
